Fixed shader loading

+++ src/graphic/Fast3D/gfx_opengl.cpp
@@ -441,7 +441,7 @@ static std::string build_vs_shader(const CCFeatures& cc_features) {
     return result;
 }
 
-static struct ShaderProgram* gfx_opengl_create_and_load_new_shader(uint64_t shader_id0, uint32_t shader_id1) {
+static struct ShaderProgram* gfx_opengl_create_and_load_new_shader(uint64_t shader_id0, uint64_t shader_id1) {
     CCFeatures cc_features;
     gfx_cc_get_features(shader_id0, shader_id1, &cc_features);
     const auto fs_buf = build_fs_shader(cc_features);
@@ -573,7 +573,7 @@ static struct ShaderProgram* gfx_opengl_create_and_load_new_shader(uint64_t shad
     return prg;
 }
 
-static struct ShaderProgram* gfx_opengl_lookup_shader(uint64_t shader_id0, uint32_t shader_id1) {
+static struct ShaderProgram* gfx_opengl_lookup_shader(uint64_t shader_id0, uint64_t shader_id1) {
     auto it = shader_program_pool.find(make_pair(shader_id0, shader_id1));
     return it == shader_program_pool.end() ? nullptr : &it->second;
 }
